% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streambugs.r
\docType{package}
\name{streambugs-package}
\alias{streambugs}
\alias{streambugs-package}
\title{Simulator of development of macroinvertebrate and algae taxa}
\description{
The main function of \pkg{streambugs} is \code{\link{run.streambugs}}.
Remaining functions have auxiliary purpose to facilitate model creation,
writing, and plotting of the simulation results, or to provide ready to use
examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.eawag.ch/en/department/siam/projects/streambugs/}
  \item Report bugs at \url{https://gitlab.com/NeleSchuwirth/streambugs/issues}
}

}
